requirements.txt
numpy == 1.17
opencv-python >= 4.1
torch == 1.7.1
torchvision
matplotlib
pycocotools
tqdm
pillow
tensorboard >= 1.14
